+++ app/controllers/Datasets.scala
@@ -567,15 +567,10 @@ class Datasets @Inject() (
           decodedCommentsByDataset += dComment
         }
 
-        val isRDFExportEnabled = current.plugin[RDFExportService].isDefined
-
-
-          filesInDataset.map
-          {
-          file =>
-            file.tags.map {
-              tag => filesTags += tag.name
-            }
+        filesInDataset.map { file =>
+          file.tags.map {
+            tag => filesTags += tag.name
+          }
         }
 
         // associated sensors
@@ -669,14 +664,16 @@ class Datasets @Inject() (
         }
         val stagingAreaDefined = play.api.Play.current.plugin[services.StagingAreaPlugin].isDefined
 
+        val extractionsByDataset = extractions.findById(new ResourceRef('dataset, id))
+
         // Increment view count for dataset
         val view_data = datasets.incrementViews(id, user)
 
         // view_data is passed as tuple in dataset case only, because template is at limit of 22 parameters
         Ok(views.html.dataset(datasetWithFiles, commentsByDataset, filteredPreviewers.toList, m,
-          decodedCollectionsInside.toList, isRDFExportEnabled, sensors, Some(decodedSpaces_canRemove), fileList,
-          filesTags, toPublish, curPubObjects, currentSpace, limit, showDownload, accessData, canAddDatasetToCollection, stagingAreaDefined,
-          view_data))
+          decodedCollectionsInside.toList, sensors, Some(decodedSpaces_canRemove), fileList,
+          filesTags, toPublish, curPubObjects, currentSpace, limit, showDownload, accessData, canAddDatasetToCollection,
+          stagingAreaDefined, view_data, extractionsByDataset))
       }
       case None => {
         Logger.error("Error getting dataset" + id)